Refactor build_job.ts and main.ts to include build job status updates
